Implemented Jump and animation
There might be a bug that I found, but cannot recreate. In the middle of jumping my character switched to IDLE state in the middle of the air, which made my character do another jump in the air.

I can't recreate it, so it may have been me not saving my script and resulting in the game running with a buggy old script, but keep an eye out for it.